refactor(basic-3.3 -> advanced-2.1 + basic-3.2): renamed to BEM approach, chenged tag use in html,js
Change all existing classes to BEM recommendations, to keep the functionality working, all usages of
previous classes were replaced in html and js files. Updated naming for HTML and CSS files, which
resulted in also accomplishing basic-3.2 and advanced-2.1. All the changes are done in accordance
with Basic and Advanced html and css recommendations.